Uploaded draft of Assignment 2
I have uploaded my draft of Assignment 2. There are a few conditions that I still need to account for while navigating the maze such as when the robot gets to the open gap in the lower part of the maze. There is another case were the robot will search one side of the start point and when backtracking, it overwrites the start point with '+'.

I will continue troubleshooting this afternoon.