Alex's final code version
This is my finalized version of code for Assignment 2. I got most features completed, except adjusting to different size mazes. In my code, I currently have to input the maze size in a pair of Row Count (RCount) and Column Count (CCount). Line 54 and 55 are the only ones that need to be changed. The rest of the code will use the updated values.